NashGames

Documentation for NashGames.

NashGames.A10a

NashGames.EPC

NashGames.EPC5nonshared

NashGames.EPC5nonsharedgreen

NashGames.EPC5shared

NashGames.EPCnonshared

NashGames.EPCshared

NashGames.NeighborsGame

NashGames.TwoGroupsSwitch

NashGames.regularEPC

NashGames.regularEPC5

NashGames.A10a — Method

Arrow and Debreu model of a competitive economy proposed in: Kenneth J. Arrow and Gerard Debreu. Existence of an equilibrium for a competitive economy. Econometrica: Journal of the Econometric Society, pages 265–290, 1954.

Also known as problem A10 in: Francisco Facchinei and Christian Kanzow. Penalty methods for the solution of general- ized Nash equilibrium problems (with complete test problems). Sapienza University of Rome, 2009.

This problem is specialized in several variants: a), b), c), d) and e)

NashGames.EPC — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

NashGames.EPC5nonshared — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() studied in *** new ***.

NashGames.EPC5nonsharedgreen — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() studied in *** new ***.

NashGames.EPC5shared — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() studied in Cojocaru, M. G., Wild, E., & Small, A. (2018). On describing the solution sets of generalized Nash games with shared constraints. Optimization and Engineering, 19(4), 845-870.

N=5

NashGames.EPCnonshared — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() studied in *** new ***.

NashGames.EPCshared — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() studied in Cojocaru, M. G., Wild, E., & Small, A. (2018). On describing the solution sets of generalized Nash games with shared constraints. Optimization and Engineering, 19(4), 845-870.

NashGames.NeighborsGame — Method

A scalable game

NashGames.TwoGroupsSwitch — Method

Inspired by an internet switching model introduced by

Alex Kesselman, Stefano Leonardi, and Vincenzo Bonifaci. Game-theoretic analysis of internet switching with selfish users. In Xiaotie Deng and Yinyu Ye, editors, Internet and Network Economics, pages 236–245, Berlin, Heidelberg, 2005. Springer Berlin Heidelberg.

and also analyzed as problem A14 in Francisco Facchinei and Christian Kanzow. Penalty methods for the solution of general- ized Nash equilibrium problems (with complete test problems). Sapienza University of Rome, 2009.

This is my* specialization.

NashGames.regularEPC — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() to find the emission of each country

NashGames.regularEPC5 — Method

Environmental protocol control proposed in Breton, M., Zaccour, G., & Zahaf, M. (2006). A game-theoretic formulation of joint implementation of environmental projects. European Journal of Operational Research, 168(1), 221-239.

Variant of EPC() to find the emission of each country